+++ test/ex_doc/formatter/html/templates_test.exs
@@ -4,7 +4,7 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
   alias ExDoc.Formatter.HTML
   alias ExDoc.Formatter.HTML.Templates
 
-  @empty_nodes_map %{modules: [], exceptions: [], protocols: []}
+  @empty_nodes_map %{modules: [], exceptions: [], protocols: [], tasks: []}
 
   defp source_url do
     "https://github.com/elixir-lang/elixir"
@@ -49,7 +49,85 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
     assert Templates.header_to_id(" ☃ ") == ""
     assert Templates.header_to_id(" &sup2; ") == ""
     assert Templates.header_to_id(" &#9180; ") == ""
-    assert Templates.header_to_id("Git Options (<code class=\"inline\">:git</code>)") == "git-options"
+    assert Templates.header_to_id("Git Options (<code class=\"inline\">:git</code>)") == "git-options-git"
+  end
+
+  test "link headers" do
+    assert Templates.link_headings("<h2>Foo</h2><h2>Bar</h2>") == """
+    <h2 id="foo" class="section-heading">
+      <a href="#foo" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Foo
+    </h2>
+    <h2 id="bar" class="section-heading">
+      <a href="#bar" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Bar
+    </h2>
+    """
+
+    assert Templates.link_headings("<h2>Foo</h2>\n<h2>Bar</h2>") == """
+    <h2 id="foo" class="section-heading">
+      <a href="#foo" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Foo
+    </h2>
+
+    <h2 id="bar" class="section-heading">
+      <a href="#bar" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Bar
+    </h2>
+    """
+
+    assert Templates.link_headings("<h2></h2><h2>Bar</h2>") == """
+    <h2></h2><h2 id="bar" class="section-heading">
+      <a href="#bar" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Bar
+    </h2>
+    """
+
+    assert Templates.link_headings("<h2></h2>\n<h2>Bar</h2>") == """
+    <h2></h2>
+    <h2 id="bar" class="section-heading">
+      <a href="#bar" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Bar
+    </h2>
+    """
+
+    assert Templates.link_headings("<h2>Foo</h2><h2></h2>") == String.trim_trailing("""
+    <h2 id="foo" class="section-heading">
+      <a href="#foo" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Foo
+    </h2>
+    <h2></h2>
+    """)
+
+    assert Templates.link_headings("<h2>Foo</h2>\n<h2></h2>") == String.trim_trailing("""
+    <h2 id="foo" class="section-heading">
+      <a href="#foo" class="hover-link"><span class="icon-link" aria-hidden="true"></span></a>
+      Foo
+    </h2>
+
+    <h2></h2>
+    """)
+  end
+
+  test "sidebar items from headers" do
+    item = %{content: nil, group: nil, id: nil, title: nil}
+    assert Templates.create_sidebar_items(%{}, [%{item | content: "<h2>Foo</h2><h2>Bar</h2>"}]) ==
+      ~s(sidebarNodes={"extras":[{"id":"","title":"","group":"","headers":[{"id":"Foo","anchor":"foo"},{"id":"Bar","anchor":"bar"}]}]})
+
+    assert Templates.create_sidebar_items(%{}, [%{item | content: "<h2>Foo</h2>\n<h2>Bar</h2>"}]) ==
+      ~s(sidebarNodes={"extras":[{"id":"","title":"","group":"","headers":[{"id":"Foo","anchor":"foo"},{"id":"Bar","anchor":"bar"}]}]})
+
+    assert Templates.create_sidebar_items(%{}, [%{item | content: "<h2></h2><h2>Bar</h2>"}]) ==
+      ~s(sidebarNodes={"extras":[{"id":"","title":"","group":"","headers":[{"id":"Bar","anchor":"bar"}]}]})
+
+    assert Templates.create_sidebar_items(%{}, [%{item | content: "<h2></h2>\n<h2>Bar</h2>"}]) ==
+      ~s(sidebarNodes={"extras":[{"id":"","title":"","group":"","headers":[{"id":"Bar","anchor":"bar"}]}]})
+
+    assert Templates.create_sidebar_items(%{}, [%{item | content: "<h2>Foo</h2><h2></h2>"}]) ==
+      ~s(sidebarNodes={"extras":[{"id":"","title":"","group":"","headers":[{"id":"Foo","anchor":"foo"}]}]})
+
+    assert Templates.create_sidebar_items(%{}, [%{item | content: "<h2>Foo</h2>\n<h2></h2>"}]) ==
+      ~s(sidebarNodes={"extras":[{"id":"","title":"","group":"","headers":[{"id":"Foo","anchor":"foo"}]}]})
   end
 
   test "synopsis" do
@@ -94,14 +172,14 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
 
   test "site title text links to homepage_url when set" do
     content = Templates.sidebar_template(doc_config(), @empty_nodes_map)
-    assert content =~ ~r{<a href="#{homepage_url()}" class="sidebar-projectLink">\n\s*<div class="sidebar-projectDetails">\n\s*<h1 class="sidebar-projectName">\n\s*Elixir\n\s*</h1>\n\s*<h2 class="sidebar-projectVersion">\n\s*v1.0.1\n\s*</h2>\n\s*</div>\n\s*</a>}
+    assert content =~ ~r{<a href="#{homepage_url()}" class="sidebar-projectLink">\s*<div class="sidebar-projectDetails">\s*<h1 class="sidebar-projectName">\s*Elixir\s*</h1>\s*<h2 class="sidebar-projectVersion">\s*v1.0.1\s*</h2>\s*</div>\s*</a>}
   end
 
   test "site title text links to main when there is no homepage_url" do
     config = %ExDoc.Config{project: "Elixir", version: "1.0.1",
                            source_root: File.cwd!, main: "hello",}
     content = Templates.sidebar_template(config, @empty_nodes_map)
-    assert content =~ ~r{<a href="hello.html" class="sidebar-projectLink">\n\s*<div class="sidebar-projectDetails">\n\s*<h1 class="sidebar-projectName">\n\s*Elixir\n\s*</h1>\n\s*<h2 class="sidebar-projectVersion">\n\s*v1.0.1\n\s*</h2>\n\s*</div>\n\s*</a>}
+    assert content =~ ~r{<a href="hello.html" class="sidebar-projectLink">\s*<div class="sidebar-projectDetails">\s*<h1 class="sidebar-projectName">\s*Elixir\s*</h1>\s*<h2 class="sidebar-projectVersion">\s*v1.0.1\s*</h2>\s*</div>\s*</a>}
   end
 
   test "list_page enables nav link when module type have at least one element" do
@@ -109,7 +187,7 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
     nodes   = ExDoc.Retriever.docs_from_modules(names, doc_config())
     modules = HTML.Autolink.all(nodes, ".html", [])
 
-    content = Templates.sidebar_template(doc_config(), %{modules: modules, exceptions: [], protocols: []})
+    content = Templates.sidebar_template(doc_config(), %{modules: modules, exceptions: [], protocols: [], tasks: []})
     assert content =~ ~r{<li><a id="modules-list" href="#full-list">Modules</a></li>}
     refute content =~ ~r{<li><a id="exceptions-list" href="#full-list">Exceptions</a></li>}
     refute content =~ ~r{<li><a id="protocols-list" href="#full-list">Protocols</a></li>}
@@ -133,11 +211,11 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
 
     # Title and headers
     assert content =~ ~r{<title>CompiledWithDocs [^<]*</title>}
-    assert content =~ ~r{<h1>\n\s*<small class="visible-xs">Elixir v1.0.1</small>\n\s*CompiledWithDocs\s*}
+    assert content =~ ~r{<h1>\s*<small class="visible-xs">Elixir v1.0.1</small>\s*CompiledWithDocs\s*}
     refute content =~ ~r{<small>module</small>}
     assert content =~ ~r{moduledoc.*Example.*CompiledWithDocs\.example.*}ms
-    assert content =~ ~r{<h2 id="module-example-unicode-escaping" class="section-heading">.*<a href="#module-example-unicode-escaping" class="hover-link">.*<i class="icon-link"></i>.*</a>.*Example.*</h2>}ms
-    assert content =~ ~r{<h3 id="module-example-h3-heading" class="section-heading">.*<a href="#module-example-h3-heading" class="hover-link">.*<i class="icon-link"></i>.*</a>.*Example H3 heading.*</h3>}ms
+    assert content =~ ~r{<h2 id="module-example-unicode-escaping" class="section-heading">.*<a href="#module-example-unicode-escaping" class="hover-link">.*<span class="icon-link" aria-hidden="true"></span>.*</a>.*Example.*</h2>}ms
+    assert content =~ ~r{<h3 id="module-example-h3-heading" class="section-heading">.*<a href="#module-example-h3-heading" class="hover-link">.*<span class="icon-link" aria-hidden="true"></span>.*</a>.*Example H3 heading.*</h3>}ms
 
     # Summaries
     assert content =~ ~r{example/2.*Some example}ms
@@ -145,13 +223,13 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
     assert content =~ ~r{example_1/0.*<span class="note">\(macro\)</span>}ms
 
     # Source
-    assert content =~ ~r{<a href="#{source_url()}/blob/master/test/fixtures/compiled_with_docs.ex#L14"[^>]*>\n\s*<i class="icon-code"></i>\n\s*</a>}ms
+    assert content =~ ~r{<a href="#{source_url()}/blob/master/test/fixtures/compiled_with_docs.ex#L14"[^>]*>\s*<span class="icon-code" aria-hidden="true"></span>\s*<span class="sr-only">View Source</span>\s*</a>}ms
 
     # Functions
     assert content =~ ~s{<div class="detail" id="example/2">}
-    assert content =~ ~s{<span id="example/1" />}
+    assert content =~ ~s{<span id="example/1"></span>}
     assert content =~ ~s{example(foo, bar \\\\ Baz)}
-    assert content =~ ~r{<a href="#example/2" class="detail-link" title="Link to this function">\n\s*<i class="icon-link"><\/i>\n\s*<\/a>}ms
+    assert content =~ ~r{<a href="#example/2" class="detail-link" title="Link to this function">\s*<span class="icon-link" aria-hidden="true"></span>\s*<span class="sr-only">Link to this function</span>\s*</a>}ms
   end
 
   test "module_page outputs the types and function specs" do
@@ -189,14 +267,14 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
 
   test "module_page outputs behavior and callbacks" do
     content = get_module_page([CustomBehaviourOne])
-    assert content =~ ~r{<h1>\n\s*<small class="visible-xs">Elixir v1.0.1</small>\n\s*CustomBehaviourOne\s*<small>behaviour</small>}m
+    assert content =~ ~r{<h1>\s*<small class="visible-xs">Elixir v1.0.1</small>\s*CustomBehaviourOne\s*<small>behaviour</small>}m
     assert content =~ ~r{Callbacks}
     assert content =~ ~r{<div class="detail" id="c:hello/1">}
     assert content =~ ~s[hello(integer)]
     assert content =~ ~s[greet(arg0)]
 
     content = get_module_page([CustomBehaviourTwo])
-    assert content =~ ~r{<h1>\n\s*<small class="visible-xs">Elixir v1.0.1</small>\n\s*CustomBehaviourTwo\s*<small>behaviour</small>\s*}m
+    assert content =~ ~r{<h1>\s*<small class="visible-xs">Elixir v1.0.1</small>\s*CustomBehaviourTwo\s*<small>behaviour</small>\s*}m
     assert content =~ ~r{Callbacks}
     assert content =~ ~r{<div class="detail" id="c:bye/1">}
   end
@@ -205,6 +283,13 @@ defmodule ExDoc.Formatter.HTML.TemplatesTest do
 
   test "module_page outputs the protocol type" do
     content = get_module_page([CustomProtocol])
-    assert content =~ ~r{<h1>\n\s*<small class="visible-xs">Elixir v1.0.1</small>\n\s*CustomProtocol\s*<small>protocol</small>\s*}m
+    assert content =~ ~r{<h1>\s*<small class="visible-xs">Elixir v1.0.1</small>\s*CustomProtocol\s*<small>protocol</small>\s*}m
+  end
+
+  ## TASKS
+
+  test "module_page outputs the task type" do
+    content = get_module_page([Mix.Tasks.TaskWithDocs])
+    assert content =~ ~r{<h1>\s*<small class="visible-xs">Elixir v1.0.1</small>\s*mix task_with_docs\s*}m
   end
 end