+++ test/ex_doc_test.exs
@@ -49,6 +49,23 @@ defmodule ExDocTest do
                  fn -> ExDoc.generate_docs project, version, options end
   end
 
+  test "default configuration" do
+    assert Map.get(%ExDoc.Config{}, :formatter) == "html"
+    assert ExDoc.Config.default(:formatter) == "html"
+
+    assert Map.get(%ExDoc.Config{}, :language) == "en"
+    assert ExDoc.Config.default(:language) == "en"
+
+    assert Map.get(%ExDoc.Config{}, :output) == "./doc"
+    assert ExDoc.Config.default(:output) == "./doc"
+
+    assert Map.get(%ExDoc.Config{}, :retriever) == ExDoc.Retriever
+    assert ExDoc.Config.default(:retriever) == ExDoc.Retriever
+
+    assert Map.get(%ExDoc.Config{}, :source_ref) == "master"
+    assert ExDoc.Config.default(:source_ref) == "master"
+  end
+
   test "version" do
     assert ExDoc.version =~ ~r{\d+\.\d+\.\d+}
   end